+++ libraries/shared/watcher/event_watcher.go
@@ -19,7 +19,6 @@ package watcher
 import (
 	"fmt"
 	"io"
-	"os"
 	"time"
 
 	"github.com/makerdao/vulcanizedb/libraries/shared/constants"
@@ -134,25 +133,6 @@ func (watcher *EventWatcher) withRetry(call func() error, expectedErrors []error
 			}
 		}
 	}
-}
-
-func addStatusForHealthCheck(msg []byte) error {
-	healthCheckFile, openErr := os.OpenFile(HealthCheckFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if openErr != nil {
-		return fmt.Errorf("error opening %s: %w", HealthCheckFile, openErr)
-	}
-	if _, writeErr := healthCheckFile.Write(msg); writeErr != nil {
-		closeErr := healthCheckFile.Close()
-		if closeErr != nil {
-			errorMsg := "error closing %s: %w -  after error writing: %s"
-			return fmt.Errorf(errorMsg, HealthCheckFile, closeErr, writeErr.Error())
-		}
-		return fmt.Errorf("error writing storage watched startup to %s: %w", HealthCheckFile, writeErr)
-	}
-	if closeErr := healthCheckFile.Close(); closeErr != nil {
-		return fmt.Errorf("error closing %s: %w", HealthCheckFile, closeErr)
-	}
-	return nil
 }
 
 func isUnexpectedError(currentError error, expectedErrors []error) bool {
